refactor: move adversarial evaluation code to src/experiments/tau-adversarial
