+++ src/experiments/tau-adversarial/data/domains/telecom/tasks_adversarial.json
