+++ src/experiments/tau-adversarial/pyproject.toml
@@ -1,0 +1,34 @@
+[project]
+name = "tau-adversarial"
+version = "0.1.0"
+description = "Adversarial evaluation module for tau2-bench - test agent robustness against manipulation attempts"
+readme = "README.md"
+requires-python = ">=3.10"
+dependencies = [
+    "tau2",
+    "loguru>=0.7.0",
+    "pytest>=8.0.0",
+]
+
+[tool.uv.sources]
+tau2 = { path = "../../..", editable = true }
+
+[project.scripts]
+tau-adversarial = "tau_adversarial.run_adversarial:main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/tau_adversarial"]
+
+[tool.ruff.lint]
+select = ["E4", "E7", "E9", "F"]
+ignore = ["E501", "F401", "F541"]
+
+[tool.ruff]
+line-length = 88
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
