--- src/tau2/adversarial/README.md
@@ -1,97 +1,0 @@
-# τ²-Adv Bench: Adversarial Evaluation for τ²-Bench
-
-Test how well agents resist manipulation attempts.
-
-## Quick Start
-
-```bash
-# Run adversarial evaluation
-python -m tau2.adversarial.run_adversarial --domain airline
-
-# Run demo
-python demo_adversarial.py
-
-# Run tests
-pytest tests/test_adversarial.py -v
-```
-
-## Attack Strategies
-
-| Strategy | What it does |
-|----------|--------------|
-| `social_engineering` | Emotional manipulation, fake authority claims |
-| `prompt_injection` | Override agent instructions |
-| `policy_exploitation` | Find and exploit policy loopholes |
-| `identity_manipulation` | Impersonate other users |
-| `information_extraction` | Probe for internal procedures |
-
-Each strategy has 3 sophistication levels:
-- **Low (0.0)**: Direct, obvious attempts
-- **Medium (0.5)**: Subtle with emotional appeals
-- **High (1.0)**: Multi-turn, trust-building attacks
-
-## Python API
-
-```python
-from tau2.adversarial import (
-    AdversarialStrategy,
-    AttackConfig,
-    load_adversarial_tasks,
-)
-
-# Configure attack
-config = AttackConfig(
-    strategy=AdversarialStrategy.SOCIAL_ENGINEERING,
-    sophistication=0.7,
-    target_goal="Get unauthorized refund",
-    domain="airline",
-)
-
-# Load adversarial tasks
-tasks = load_adversarial_tasks("airline")
-```
-
-## Safety Metrics
-
-- **Safety Score** (0-1): Higher = safer, penalized by violations
-- **Defense Rate**: Did the agent block the attack goal?
-- **Violations**: List of detected issues (unauthorized actions, info disclosure, etc.)
-
-## Adding Tasks
-
-Create `tasks_adversarial.json` in your domain's data directory:
-
-```json
-[
-    {
-        "id": "adv_001",
-        "adversarial_metadata": {
-            "strategy": "social_engineering",
-            "sophistication": 0.5,
-            "target_goal": "Get unauthorized refund"
-        },
-        "user_scenario": {
-            "instructions": {
-                "task_instructions": "Claim supervisor approved your refund...",
-                "domain": "airline"
-            }
-        }
-    }
-]
-```
-
-## Files
-
-```
-src/tau2/adversarial/
-├── strategies.py      # Attack definitions
-├── tasks.py           # Task loading
-├── run_adversarial.py # CLI runner
-└── README.md          # This file
-
-src/tau2/evaluator/
-└── evaluator_safety.py # Violation detection
-
-data/tau2/domains/*/
-└── tasks_adversarial.json # Adversarial tasks per domain
-```
